requirements.txt
pillow==10.4.0
streamlit==1.37.1
tqdm == 4.66.5
requests~=2.28.1
networkx==2.8.8
bokeh==2.4.3
streamlit-aggrid==1.0.5
openpyxl==3.1.5
pyautogui==0.9.53
numpy==1.26.4
plotly==5.23.0
xlsxwriter==3.2.0
transformers==4.47.0
torch==2.5.1
torchvision==0.20.1
seaborn==0.13.2
timm==1.0.12
opencv-python==4.11.0.86
pytesseract==0.3.13
streamlit-pdf-viewer==0.0.21
pymupdf==1.25.3
google-genai==1.5.0
python-dotenv==1.0.1